// projects/ccf-database/src/lib/interfaces.d.ts
export interface ListResultItem {
    /** Identifier */
    '@id': string;
    /** Label */
    label: string;
    /** Description */
    description: string;
    /** Associated URL */
    link: string;
}
/** Donor List Result */
export interface DonorResult extends ListResultItem {
    /** JSON-LD Type */
    '@type': 'Donor';
    /** Provider Name */
    providerName: string;
}
/** Dataset List Result */
export interface DatasetResult extends ListResultItem {
    /** JSON-LD Type */
    '@type': 'Dataset';
    /** Technology used in data */
    technology: string;
    /** Representative thumbnail URL */
    thumbnail: string;
}
/** Tissue Section List Result */
export interface TissueSectionResult extends ListResultItem {
    /** JSON-LD Type */
    '@type': 'Sample';
    /** Sample tissue section type */
    sampleType: 'Tissue Section' | 'Non-Standard';
    /** Index into the sequence of sections from the parent block */
    sectionNumber: number;
    /** Datasets derived from this tissue section */
    datasets: DatasetResult[];
}
/** Tissue Block List Result */
export interface TissueBlockResult extends ListResultItem {
    /** JSON-LD Type */
    '@type': 'Sample';
    /** Sample tissue block type */
    sampleType: 'Tissue Block' | 'Non-Standard';
    /** Number of sections extracted from this block */
    sectionCount: number;
    /** Size of each tissue section */
    sectionSize: number;
    /** Units of the section size (generally, millimeters) */
    sectionUnits: string;
    /** The donor this tissue block was derived from */
    donor: DonorResult;
    /** The associated spatial entity (rui_location) for this block */
    spatialEntityId: string;
    /** The sections extracted from this tissue block */
    sections: TissueSectionResult[];
    /** Datasets derived from this tissue block */
    datasets: DatasetResult[];
}